[#58] Refactor: 캘린더 재활용 가능하도록 설계 및 ScheduleList 분리
